Restructured files & code.
